add an endpoint
