+++ swagger.yml
@@ -8,27 +8,25 @@ consumes:
 produces:
   - "application/json"
 
-basePath: "/api"
+basePath: "/2016-09-06"
 
 # Paths supported by the server application
 paths:
-  /people:
+  /sequences:
     get:
-      operationId: "people.read"
+      operationId: "sequences.read"
       tags:
-        - "People"
-      summary: "The people data structure supported by the server application"
-      description: "Read the list of people"
+        - "Sequences"
+      summary: "The sequence data strucutre"
+      description: "Read attributes of sequence"
       responses:
         200:
-          description: "Successful read people list operation"
+          description: "Successful read sequence operation"
           schema:
             type: "array"
             items:
               properties:
-                fname:
-                  type: "string"
-                lname:
-                  type: "string"
-                timestamp:
-                  type: "string"
+                  seguid:
+                    type: "string"
+                  timestamp:
+                    type: "string"
